requirements.txt
pandas
numpy
sqlalchemy
python-dotenv
psycopg2-binary
openpyxl
